whoops wrong data type --fixed

+++ src/main/java/com/team766/ViSIONbase/GrayScaleCamera.java
@@ -61,7 +61,7 @@ public class GrayScaleCamera extends PhotonCamera {
         }
     }
 
-    public PhotonTrackedTarget getTrackedTargetWithID(double ID) throws AprilTagGeneralCheckedException {
+    public PhotonTrackedTarget getTrackedTargetWithID(int ID) throws AprilTagGeneralCheckedException {
         var result = getLatestResult();
         boolean hasTargets =
                 result.hasTargets();
